requirements.txt
streamlit
opencv-python-headless
numpy
joblib
Pillow
google-generativeai
requests
scikit-learn
plotly 
python-dotenv
